+++ frontend/containers/DocumentListToolbar/DocumentListToolbar.jsx
@@ -12,7 +12,6 @@ import * as documentActions from 'actions/documentActions'
 import * as documentsActions from 'actions/documentsActions'
 
 import {bindActionCreators} from 'redux'
-import {buildUrl} from 'lib/router'
 import {connectHelper} from 'lib/util'
 import {Format} from 'lib/util/string'
 import {route} from '@dadi/preact-router'
@@ -193,11 +192,10 @@ class DocumentListToolbar extends Component {
       referencedField
     } = this.props
 
-    if (referencedField) {
-      return buildUrl(...onBuildBaseUrl(), 'select', referencedField, page)
-    }
-
-    return buildUrl(...onBuildBaseUrl(), page)
+    return onBuildBaseUrl({
+      page: page,
+      referenceFieldSelect: referencedField
+    })
   }
 
   handleBulkActionApply(actionType) {
@@ -282,20 +280,19 @@ class DocumentListToolbar extends Component {
       group
     })
 
-    const referenceFieldSchema = currentParentCollection.fields[referencedField]
-    const referenceFieldSection = referenceFieldSchema &&
-      referenceFieldSchema.publish &&
-      referenceFieldSchema.publish.section &&
-      Format.slugify(referenceFieldSchema.publish.section)
+    let redirectUrl = onBuildBaseUrl({
+      createNew: !Boolean(state.router.parameters.documentId)
+    })
 
-    route(buildUrl(...onBuildBaseUrl({section: referenceFieldSection})))
+    route(redirectUrl)
   }
 }
 
 export default connectHelper(
   state => ({
     api: state.api,
-    documents: state.documents
+    documents: state.documents,
+    router: state.router
   }),
   dispatch => bindActionCreators({
     ...appActions,