+++ frontend/containers/ProfileEdit/ProfileEdit.jsx
@@ -32,12 +32,6 @@ class ProfileEdit extends Component {
     onBuildBaseUrl: proptypes.func,
 
     /**
-    * A callback to be used to obtain the sibling document routes (edit, create and list), as
-    * determined by the view.
-    */
-    onGetRoutes: proptypes.func,
-
-    /**
      * The name of a reference field currently being edited.
      */
     referencedField: proptypes.string,
@@ -74,7 +68,6 @@ class ProfileEdit extends Component {
   render() {
     const {
       onBuildBaseUrl,
-      onGetRoutes,
       referencedField,
       section,
       state
@@ -142,12 +135,13 @@ class ProfileEdit extends Component {
     if (field === 'clientId') {
       return (
         <FieldString
+          displayName="Username"
           error={error}
+          name="clientId"
           onChange={this.handleFieldChange.bind(this)}
           onError={this.handleFieldError.bind(this)}
           schema={{
             ...schema,
-            label: 'Username',
             publish: {
               readonly: true
             }
@@ -160,13 +154,12 @@ class ProfileEdit extends Component {
     if (field === 'secret') {
       return (
         <FieldPassword
+          displayName="Password"
           error={error}
+          name="secret"
           onChange={this.handleFieldChange.bind(this)}
           onError={this.handleFieldError.bind(this)}
-          schema={{
-            ...schema,
-            label: 'Password'
-          }}
+          schema={schema}
         />
       )
     }
@@ -174,13 +167,12 @@ class ProfileEdit extends Component {
     if (field === 'data.publishFirstName') {
       return (
         <FieldString
+          displayName="First name"
           error={error}
+          name="data.publishFirstName"
           onChange={this.handleFieldChange.bind(this)}
           onError={this.handleFieldError.bind(this)}
-          schema={{
-            ...schema,
-            label: 'First name'
-          }}
+          schema={schema}
           value={userData.publishFirstName}
         />
       )
@@ -189,13 +181,12 @@ class ProfileEdit extends Component {
     if (field === 'data.publishLastName') {
       return (
         <FieldString
+          displayName="Last name"
           error={error}
+          name="data.publishLastName"
           onChange={this.handleFieldChange.bind(this)}
           onError={this.handleFieldError.bind(this)}
-          schema={{
-            ...schema,
-            label: 'Last name'
-          }}
+          schema={schema}
           value={userData.publishLastName}
         />
       )